리액트 테스트

+++ src/main/java/com/kh/member_project/domain/Board.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class Board {
-    private int bno;
+    private long bno;
     private String  title;
     private String content;
     private String writer;
