+++ src/main/java/com/kh/member_project/service/BoardService/BoardService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface BoardService {
     void register(Board board);
     List<Board> boardList(String keyword);
-    Board readBoard(int bno);
-    void deleteBoard(int bno);
+    Board readBoard(long bno);
+    void deleteBoard(long bno);
     void updateBoard(Board board);
 }
